Refactor dict_zip type stubs and module structure
- Dropped support for Python versions earlier than 3.9.
- Removed separate .pyi stub files; type hints are now integrated directly in .py files.
- Split dict_zip and dict_zip_longest into separate modules for clarity and modularity.
